requirements.txt
blinker==1.9.0
click==8.3.0
dotenv==0.9.9
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.3.4
pandas==2.3.3
pyodbc==5.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
tzdata==2025.2
Werkzeug==3.1.3
langchain
langchain-openai
tqdm
tiktoken
mlflow
uvicorn
pytest
Pyyaml
dvc
